requirements.txt
streamlit>=1.29.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
faiss-cpu>=1.8.0
GitPython>=3.1.40
ollama>=0.1.6
sentence-transformers>=2.2.2
torch>=2.0.0
